requirements.txt
local-attention=1.4.3
numpy=1.21.5
pandas=1.2.4
python=3.8.13
pytorch=1.12.1
scanpy=1.9.1
scikit-learn=1.1.1
scipy=1.6.2